agent/pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "AI agent that uses MCP server tools via OpenAI"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "openai",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
